Added n-gram termination check
Main calls to all `VerbTense` subclasses now check the last word in the candidate n-gram to ensure that the last word in the sequence matches the expected part-of-speech (e.g. VBG for continuous tenses).

This is necessary to ensure that matched n-grams do not span more words than appropriate for the tense.